refactor: adjust jwt strategy
