+++ package.json
@@ -35,6 +35,7 @@
     "@nestjs/bull": "10.0.1",
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
+    "@nestjs/jwt": "10.1.0",
     "@nestjs/passport": "10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@prisma/client": "5.1.1",
