+++ src/common/infra/http/auth/auh.module.ts
@@ -1,0 +1,18 @@
+import { Module } from '@nestjs/common';
+import { APP_GUARD } from '@nestjs/core';
+import { PassportModule } from '@nestjs/passport';
+
+import { JWTAuthGuard } from './jwt-auth-guard';
+import { JwtStrategy } from './jwt.strategy';
+
+@Module({
+  imports: [PassportModule],
+  providers: [
+    JwtStrategy,
+    {
+      provide: APP_GUARD,
+      useClass: JWTAuthGuard,
+    },
+  ],
+})
+export class AuthModule {}
