+++ src/common/infra/http/auth/decorators/current-user.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 
-import { AuthUser } from './auth-user';
+import { AuthUser } from '../auth-user';
 
 export const CurrentUser = createParamDecorator(
   (data: unknown, context: ExecutionContext): AuthUser => {
