+++ src/modules/timeline/infra/http/controllers/project.controller.ts
@@ -1,4 +1,3 @@
-import { IsPublic } from '@common/infra/http/auth/is-public';
 import { CreateProjectUseCase } from '@modules/timeline/application/use-cases/create-project';
 import { Body, Controller, Post } from '@nestjs/common';
 
@@ -9,7 +8,6 @@ export class ProjectController {
   constructor(private readonly createProjectUseCase: CreateProjectUseCase) {}
 
   @Post()
-  @IsPublic()
   async createProject(@Body() body: CreateProjectDTO) {
     const { authorId, content, requirements, roles, technologies, title } =
       body;
